Unify DBO yield helpers and drop compile guard
